Minicute.Data.GMachine.Instruction

Description

Instructions and other helper types for G-Machine

G-Machine Architecture

Abstract Structure

This G-Machine is represented by 6-tuple.

\((CurrentCode,\ AddressStack,\ ValueStack,\ NodeHeap,\ GlobalEnvironment)\)

CurrentCode

a list of codes currently executed.

AddressStack

a stack of G-Machine addresses.

ValueStack

a stack of primitive values.

EvalDump

a stack of snapshots of codes, address stacks, and value stacks.

NodeHeap

a heap of G-Machine nodes.

GlobalEnvironment

a map from identifiers to addresses of corresponding nodes.

The following tuple describes the initial state for the machine.

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & [\text{IMakeGlobal}\ \unicode{x201C}main\unicode{x201D}, \text{IEval}], & [], & [], & [], & initialHeap, & initialGlobal & ) \end{array}\\ & \text{(when $initialHeap$ is a heap including all NGlobals for the supercombinators in a program)}\\ & \text{(when $initialGlobal$ is a global including all supercombinators in a program)} \end{align} \]

TODO: make this G-Machine be parallel

Operational Semantics

Any unspecified cases mean erratic state.

Basic Node Creating Operations

IMakeInteger

\[ \begin{array}{r r r r r l l l} ( & \text{IMakeInteger}\ n : codes, & addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr : addrs, & values, & dump, & heap[addr: \text{NInteger}\ n], & global & ) \end{array} \]

IMakeConstructor

\[ \begin{array}{r r r r r l l l} ( & \text{IMakeConstructor}\ t\ n : codes, & addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr : addrs, & values, & dump, & heap[addr: \text{NConstructor}\ t\ n], & global & ) \end{array} \]

IMakeStructure

\[ \begin{array}{r r r r r l l l} ( & \text{IMakeStructure}\ t\ n : codes, & addr_0 : addr_1 : ... : addr_{n - 1} : addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr : addrs, & values, & dump, & heap[addr: \text{NStructure}\ t\ [addr_{n - 1}, ..., addr_1, addr_0]], & global & ) \end{array} \]

IMakeApplication

\[ \begin{array}{r r r r r l l l} ( & \text{IMakeApplication} : codes, & addr_0 : addr_1 : addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr : addrs, & values, & dump, & heap[addr: \text{NApplication}\ addr_1\ addr_0], & global & ) \end{array} \]

IMakeGlobal

\[ \begin{array}{r r r r r l l l} ( & \text{IMakeGlobal}\ id : codes, & addrs, & values, & dump, & heap, & global[id: addr] & )\\ \hline ( & codes, & addr : addrs, & values, & dump, & heap, & global[id: addr] & ) \end{array} \]

IMakePlaceholders

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & \text{IMakePlaceholders}\ n : codes, & addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap', & global & ) \end{array}\\ & heap' = heap[addr_0: \text{NEmpty}, addr_1: \text{NEmpty}, ..., addr_n: \text{NEmpty}] \end{align} \]

Do we need to add NEmpty or just use NIndirect nullAddr?

Address Stack Based Operations

IPop

\[ \begin{array}{r r r r r l l l} ( & \text{IPop}\ n : codes, & addr_0 : addr_1 : ... : addr_{n - 1} : addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addrs, & values, & dump, & heap, & global & ) \end{array} \]

IDig

\[ \begin{array}{r r r r r l l l} ( & \text{IDig}\ n : codes, & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr_0 : addrs, & values, & dump, & heap, & global & ) \end{array} \]

IUpdate

\[ \begin{array}{r r r r r l l l} ( & \text{IUpdate}\ n : codes, & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap[addr_n: \text{NIndirect}\ addr_0], & global & ) \end{array} \]

ICopy

\[ \begin{array}{r r r r r l l l} ( & \text{ICopy}\ n : codes, & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addr_n : addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap, & global & ) \end{array} \]

Value Stack Based Operations

IPushBasicValue

\[ \begin{array}{r r r r r l l l} ( & \text{IPushBasicValue}\ v : codes, & addrs, & values, & dump, & heap, & global & )\\ \hline ( & codes, & addrs, & v : values, & dump, & heap, & global & ) \end{array} \]

IPushExtractedValue

\[ \begin{array}{r r r r r l l l} ( & \text{IPushExtractedValue} : codes, & addr : addrs, & values, & dump, & heap[addr: \text{NInteger}\ v], & global & )\\ \hline ( & codes, & addrs, & v : values, & dump, & heap, & global & ) \end{array} \]

\[ \begin{array}{r r r r r l l l} ( & \text{IPushExtractedValue} : codes, & addr : addrs, & values, & dump, & heap[addr: \text{NStructure}\ v\ []], & global & )\\ \hline ( & codes, & addrs, & v : values, & dump, & heap, & global & ) \end{array} \]

IWrapAsInteger

\[ \begin{array}{r r r r r l l l} ( & \text{IWrapAsInteger} : codes, & addrs, & v : values, & dump, & heap, & global & )\\ \hline ( & codes, & addr : addrs, & values, & dump, & heap[addr: \text{NInteger}\ v], & global & ) \end{array} \]

IWrapAsStructure

\[ \begin{array}{r r r r r l l l} ( & \text{IWrapAsStructure} : codes, & addrs, & v : values, & dump, & heap, & global & )\\ \hline ( & codes, & addr : addrs, & values, & dump, & heap[addr: \text{NStructure}\ v\ []], & global & ) \end{array} \]

IUpdateAsInteger

\[ \begin{array}{r r r r r l l l} ( & \text{IUpdateAsInteger}\ n : codes, & addr_0 : addr_1 : ... : addr_n : addrs, & v : values, & dump, & heap, & global & )\\ \hline ( & codes, & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap[addr_n: \text{NInteger}\ v], & global & ) \end{array} \]

IUpdateAsStructure

\[ \begin{array}{r r r r r l l l} ( & \text{IUpdateAsStructure}\ n : codes, & addr_0 : addr_1 : ... : addr_n : addrs, & v : values, & dump, & heap, & global & )\\ \hline ( & codes, & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap[addr_n: \text{NStructure}\ v\ []], & global & ) \end{array} \]

Primitive Operations

IPrimitive

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & \text{IPrimitive}\ op : codes, & addrs, & v_0 : v_1 : values, & dump, & heap, & global & )\\ \hline ( & codes, & addrs, & v' : values, & dump, & heap, & global & ) \end{array}\\ & v' = v_0\ R\ v1\\ & \text{(when $op$ represents a binary operation $R$)} \end{align} \]

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & \text{IPrimitive}\ op : codes, & addrs, & v : values, & dump, & heap, & global & )\\ \hline ( & codes, & addrs, & v' : values, & dump, & heap, & global & ) \end{array}\\ & v' = R\ v\\ & \text{(when $op$ represents a unary operation $R$)} \end{align} \]

Node Inspecting Operations

IUnwind

\[ \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr : addrs, & values, & (codes', addrs', values') : dump, & heap[addr: \text{NInteger}\ n], & global & )\\ \hline ( & codes', & addr : addrs', & values', & dump, & heap, & global & ) \end{array} \]

\[ \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr : addrs, & values, & (codes', addrs', values') : dump, & heap[addr: \text{NStructure}\ t\ fAddrs], & global & )\\ \hline ( & codes', & addr : addrs', & values', & dump, & heap, & global & ) \end{array} \]

\[ \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr : addrs, & values, & dump, & heap[addr: \text{NApplication}\ addr_1\ addr_0], & global & )\\ \hline ( & [\text{IUnwind}], & addr_1 : addr : addrs, & values, & dump, & heap, & global & ) \end{array} \]

\[ \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr : addrs, & values, & dump, & heap[addr: \text{NIndirect}\ addr'], & global & )\\ \hline ( & [\text{IUnwind}], & addr' : addrs, & values, & dump, & heap, & global & ) \end{array} \]

\[ \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr : addrs, & values, & dump, & heap[addr: \text{NConstructor}\ t\ 0], & global & )\\ \hline ( & \text{constructorCode}\ t, & addr : addrs, & values, & dump, & heap, & global & ) \end{array} \]

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap[addr_0: \text{NConstructor}\ t\ n], & global & )\\ \hline ( & \text{IRearrange}\ n : \text{constructorCode}\ t, & addr_1 : ... : addr_n : addrs, & values, & dump, & heap, & global & ) \end{array}\\ & \text{(when $n > 0$)} \end{align} \]

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & [addr_0, addr_1, ..., addr_m], & values, & dump, & heap[addr_0: \text{NConstructor}\ t\ n], & global & )\\ \hline ( & [\text{Return}], & [addr_0, addr_1, ..., addr_m], & values, & dump, & heap, & global & ) \end{array}\\ & \text{(when $m < n$)} \end{align} \]

\[ \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr : addrs, & values, & dump, & heap[addr: \text{NGlobal}\ 0\ codes'], & global & )\\ \hline ( & codes', & addr : addrs, & values, & dump, & heap, & global & ) \end{array} \]

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & addr_0 : addr_1 : ... : addr_n : addrs, & values, & dump, & heap[addr_0: \text{NGlobal}\ n\ codes'], & global & )\\ \hline ( & \text{IRearrange}\ n : codes', & addr_1 : ... : addr_n : addrs, & values, & dump, & heap, & global & ) \end{array}\\ & \text{(when $n > 0$)} \end{align} \]

\[ \begin{align} & \begin{array}{r r r r r l l l} ( & [\text{IUnwind}], & [addr_0, addr_1, ..., addr_m], & values, & dump, & heap[addr_0: \text{NGlobal}\ t\ codes'], & global & )\\ \hline ( & [\text{Return}], & [addr_0, addr_1, ..., addr_m], & values, & dump, & heap, & global & ) \end{array}\\ & \text{(when $m < n$)} \end{align} \]

IDestruct

\[ \begin{array}{r r r r r l l l} ( & \text{IDestruct}\ n : codes, & addr : addrs, & values, & dump, & heap[addr: \text{NStructure}\ t\ [addr_{n - 1}, ..., addr_1, addr_0]], & global & )\\ \hline ( & codes, & addr_0 : addr_1 : ... : addr_{n - 1} : addrs, & values, & dump, & heap, & global & ) \end{array} \]

Dump Related Operations

IEval

\[ \begin{array}{r r r r r l l l} ( & \text{IEval} : codes, & addr : addrs, & values, & dump, & heap, & global & )\\ \hline ( & [\text{IUnwind}], & [addr], & [], & (codes, addrs, values) : dump, & heap, & global & ) \end{array} \]

IReturn

\[ \begin{array}{r r r r r l l l} ( & [\text{IReturn}], & [addr_0, addr_1, ..., addr_n], & values, & (codes', addrs', values') : dump, & heap, & global & )\\ \hline ( & codes', & addr_n : addrs', & values', & dump, & heap, & global & ) \end{array} \]

Match Operations

IMatch

\[ \begin{array}{r r r r r l l l} ( & \text{IMatch}\ table[t: casecode_t] : codes, & addr : addrs, & values, & dump, & heap[addr: \text{NStructure}\ t\ fAddrs], & global & )\\ \hline ( & casecode_t \diamond codes, & addr : addrs, & values, & dump, & heap, & global & ) \end{array} \]

Virtual Operations

Following operations are not real constructors of Instruction. These exist only for semantic purposes.

IRearrange

\[ \begin{array}{r r r r r l l l} ( & \text{IRearrange}\ n : codes, & addr_0 : addr_1 : ... : addr_{n - 1} : addrs, & values, & dump, & heap[addr_i: \text{NApplication}\ addr'_i\ addr''_i], & global & )\\ \hline ( & codes, & addr''_0 : addr''_1 : ... : addr''_{n - 1} : addr_{n - 1} : addrs, & values, & dump, & heap, & global & ) \end{array}\\ \]

Types

type GMachineProgram = [GMachineSupercombinator]

A G-Machine program.

type GMachineSupercombinator = (Identifier, Int, GMachineExpression)

A G-Machine top-level definition.

It contains

Identifier

The identifier for the Supercombinator.

Int

The arity of the Supercombinator.

GMachineExpression

The code for the Supercombinator.

type GMachineExpression = [Instruction]

A G-Machine expression

Instructions

data Instruction

A G-Machine instruction

Constructors

IMakeInteger Integer

IMakeConstructor Integer Integer

IMakeStructure Integer Integer

IMakeApplication

IMakeGlobal Identifier

IMakePlaceholders Int

IPop Int

IDig Int

IUpdate Int

ICopy Int

IPushBasicValue Integer

IPushExtractedValue

IWrapAsInteger

IWrapAsStructure

IUpdateAsInteger Int

IUpdateAsStructure Int

IPrimitive Primitive

IUnwind

IDestruct Integer

IEval

IReturn

IMatch MatchTable

newtype MatchTable

Table expressing a match expression.

[MatchEntry]

cases in the match expression.

Constructors

MatchTable [MatchEntry]

newtype MatchEntry

Entry expressing a match case.

Integer

the tag of the case.

GMachineExpression

the instructions of the case.

Constructors

MatchEntry (Integer, GMachineExpression)